[Feat]: 인증/인가 로직 적용

+++ src/main/java/com/backend/domain/product/controller/ApiV1ProductController.java
@@ -2,6 +2,7 @@ package com.backend.domain.product.controller;
 
 import com.backend.domain.member.entity.Member;
 import com.backend.domain.member.repository.MemberRepository;
+import com.backend.domain.member.service.MemberService;
 import com.backend.domain.product.dto.*;
 import com.backend.domain.product.entity.Product;
 import com.backend.domain.product.enums.AuctionStatus;
@@ -15,6 +16,8 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -26,15 +29,16 @@ import java.util.List;
 public class ApiV1ProductController implements ApiV1ProductControllerDocs {
     private final ProductService productService;
     private final MemberRepository memberRepository;
+    private final MemberService memberService;
 
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     @Transactional
     public RsData<ProductDto> createProduct(
             @RequestPart("request") @Valid ProductCreateRequest request,
-            @RequestPart("images") List<MultipartFile> images
-//            @AuthenticationPrincipal Member actor
+            @RequestPart("images") List<MultipartFile> images,
+            @AuthenticationPrincipal User user
     ) {
-        Member actor = memberRepository.findAll().getFirst();
+        Member actor = memberService.findMemberByEmail(user.getUsername());
 
         Product product = productService.create(actor, request, images);
 
@@ -83,12 +87,13 @@ public class ApiV1ProductController implements ApiV1ProductControllerDocs {
             @PathVariable Long productId,
             @RequestPart("request") @Valid ProductModifyRequest request,
             @RequestPart(value = "images", required = false) List<MultipartFile> images,
-            @RequestPart(value = "deleteImageIds", required = false) List<Long> deleteImageIds
-//            @AuthenticationPrincipal Member actor
+            @RequestPart(value = "deleteImageIds", required = false) List<Long> deleteImageIds,
+            @AuthenticationPrincipal User user
     ) {
+        Member actor = memberService.findMemberByEmail(user.getUsername());
         Product product = productService.getProductById(productId);
 
-//        product.checkActorCanModify(actor);
+        product.checkActorCanModify(actor);
 
         productService.modifyProduct(product, request, images, deleteImageIds);
 
@@ -102,12 +107,13 @@ public class ApiV1ProductController implements ApiV1ProductControllerDocs {
     @DeleteMapping("/{productId}")
     @Transactional
     public RsData<Void> deleteProduct(
-            @PathVariable Long productId
-//            @AuthenticationPrincipal Member actor
+            @PathVariable Long productId,
+            @AuthenticationPrincipal User user
     ) {
+        Member actor = memberService.findMemberByEmail(user.getUsername());
         Product product = productService.getProductById(productId);
 
-//        product.checkActorCanDelete(actor);
+        product.checkActorCanDelete(actor);
 
         productService.deleteProduct(product);
 
@@ -120,10 +126,10 @@ public class ApiV1ProductController implements ApiV1ProductControllerDocs {
             @RequestParam(defaultValue = "1") int page,
             @RequestParam(defaultValue = "20") int size,
             @RequestParam(defaultValue = "SELLING") SaleStatus status,
-            @RequestParam(defaultValue = "LATEST") ProductSearchSortType sort
-//            @AuthenticationPrincipal Member actor
+            @RequestParam(defaultValue = "LATEST") ProductSearchSortType sort,
+            @AuthenticationPrincipal User user
     ) {
-        Member actor = memberRepository.findAll().getFirst();
+        Member actor = memberService.findMemberByEmail(user.getUsername());
 
         Page<Product> products = productService.findByMemberPaged(page, size, sort, actor, status);
